More Highscore Fixes
Handles adding player scores properly when there are fewer than 10
highscores.

+++ Assets/Scripts/Global.cs
@@ -76,7 +76,7 @@ public class Global	 : MonoBehaviour
 			break;
 		case GameState.GAME_OVER:
 			// GAME OVER
-			PlayerPrefs.SetFloat("player score",ScoreScript.getScore());
+			PlayerPrefs.SetInt("player score",(int) ScoreScript.getScore());
 			Application.LoadLevel("HighScoreMenu");
 			break;
 		default:
